fix: fetch and use group id over group name from config
Users can define the group identifier `config["group"]` as either a:
* Group Name
* Group Username
* Group ID

of which the Group Name and the Username are mutable and can be changed
while tgarchive is running, causing `get_messages()` to fail. For such
cases, we can always ensure that Group IDs are being used to synchronize
messages from the Telegram API.

This commit makes the required internal changes for fetching the canonical
Group ID from `config["group"]` and using it in `get_messages()`. The end-users
can continue using their existing configuration without any changes.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>